+++ backend/goodday/src/main/java/seeuthere/goodday/location/domain/location/WeightStations.java
@@ -1,7 +1,6 @@
 package seeuthere.goodday.location.domain.location;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
