+++ backend/goodday/src/main/java/seeuthere/goodday/path/repository/SubwayRedisRepository.java
@@ -1,7 +1,7 @@
 package seeuthere.goodday.path.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import seeuthere.goodday.path.service.TransportCache;
+import seeuthere.goodday.path.domain.TransportCache;
 
 public interface SubwayRedisRepository extends CrudRepository<TransportCache, String> {
